saving progress. Moved all funcs out of chorelistview
