+++ Server/Model/HouseHold.cs
@@ -3,13 +3,8 @@ using DTO;
 
 namespace Model;
 
-
-
 public class Household
 {
-    public Household(){
-        
-    }
     public int Id { get; set; }
     public string Name { get; set; } = "";
     public int Code { get; set; }
@@ -19,7 +14,7 @@ public class Household
     public List<int> AvailableAvatars { get; set; } = new List<int>();
     
     public ICollection<Profile> Profiles { get; }  = new List<Profile>();
-
+    public ICollection<Chore> Chores { get; }  = new List<Chore>();
 
     public Household(HouseholdDto dto)
     {    
@@ -27,5 +22,10 @@ public class Household
         UserId = dto.OwnerId;
         Random random = new Random();
         Code = random.Next(1000, 9999);        
+    }
+
+    public Household()
+    {
+        
     }
 }
